Move createInteractions and related features to a separate folder
Relates: #80, #33
