+++ packages/plugins/Draw/src/store/createInteractions/createDeleteInteractions.ts
