+++ packages/plugins/Draw/src/store/createInteractions/createTextInteractions.ts
@@ -2,8 +2,8 @@ import { Draw, Snap } from 'ol/interaction'
 import Interaction from 'ol/interaction/Interaction'
 import VectorSource from 'ol/source/Vector'
 import { PolarActionContext } from '@polar/lib-custom-types'
-import { DrawGetters, DrawState } from '../types'
-import { createTextStyle } from '../utils/createTextStyle'
+import { DrawGetters, DrawState } from '../../types'
+import { createTextStyle } from '../../utils/createTextStyle'
 
 export default function (
   {
